Restructure according to namespaces
