+++ BUILD
@@ -2,9 +2,12 @@ load("@rules_cc//cc:defs.bzl", "cc_library")
 
 cc_library(
     name = "habitify_event_bus",
-    hdrs = [
-        "include/habitify_event.h",
-        "include/habitify_event_bus.h",
+    srcs = [
+        "src/event_bus_impl.cc",
+        "src/listener.cc",
+        "src/port.cc",
+        "src/publisher.cc",
     ],
+    includes = ["include"],
     visibility = ["//visibility:public"],
 )
